big improvement ~kasper
