+++ CMakeLists.txt
@@ -3,8 +3,8 @@ project(netpbm LANGUAGES C)
 
 set(CMAKE_C_STANDARD 23)
 
-set(SOURCE_FILES main.c ppm.h pgm.h pbm.h bayer.h)
-set(HEADER_FILES ppm.h pgm.h pbm.h bayer.h)
+set(SOURCE_FILES main.c ppm.h pgm.h pbm.h bayer.h types/ppm.h types/pgm.h types/pbm.h types/pixel.h)
+set(HEADER_FILES ppm.h pgm.h pbm.h bayer.h types/ppm.h types/pgm.h types/pbm.h types/pixel.h)
 set_source_files_properties(${SOURCE_FILES} PROPERTIES LANGUAGE C)
 set_source_files_properties(${HEADER_FILES} PROPERTIES HEADER_FILE_ONLY TRUE)
