+++ bayer.h
@@ -1,6 +1,14 @@
 #ifndef NETPBM__BAYER_H_
 #define NETPBM__BAYER_H_
 
+#include <stdint.h>
+
+typedef struct {
+  uint32_t width_;
+  uint32_t height_;
+  uint8_t *data_;
+} BayerImage;
+
 const double kBayer2X2[2][2] = {
 	{0.0 / 4.0, 2.0 / 4.0},
 	{3.0 / 4.0, 1.0 / 4.0}
