+++ ppm.h
@@ -6,25 +6,12 @@
 #include <stdlib.h>
 #include <stdint.h>
 #include <stdio.h>
-#include "pgm.h"
+#include <string.h>
+#include "types/ppm.h"
+#include "types/pgm.h"
 
 #define PPM_MAX_COLOR 255
 #define PPM_MAX_COLOR_F 255.0
-
-// Pixel struct
-typedef struct {
-  uint8_t r_;
-  uint8_t g_;
-  uint8_t b_;
-} Pixel;
-
-// PPM image format
-typedef struct {
-  uint32_t width_;
-  uint32_t height_;
-  uint16_t max_color_;
-  Pixel *data_;
-} PpmImage;
 
 /**
  * Allocate memory for a PPM image.
@@ -35,19 +22,19 @@ typedef struct {
  */
 PpmImage *AllocatePpm(uint32_t width, uint32_t height) {
   // Allocate memory for image data
-  PpmImage *image = (PpmImage *)malloc(sizeof(PpmImage));
+  PpmImage *image = (PpmImage *) malloc(sizeof(PpmImage));
   if (!image) {
-	fprintf(stderr, "Error: out of memory\n");
-	return NULL;
+    fprintf(stderr, "Error: out of memory\n");
+    return NULL;
   }
   image->width_ = width;
   image->height_ = height;
   image->max_color_ = PPM_MAX_COLOR;
-  image->data_ = (Pixel *)malloc(width * height * sizeof(Pixel));
+  image->data_ = (Pixel *) malloc(width * height * sizeof(Pixel));
   if (!image->data_) {
-	fprintf(stderr, "Error: out of memory\n");
-	free(image);
-	return NULL;
+    fprintf(stderr, "Error: out of memory\n");
+    free(image);
+    return NULL;
   }
 
   return image;
@@ -63,8 +50,8 @@ PpmImage *ReadPpm(const char *filename) {
   // Open file for reading
   FILE *fp = fopen(filename, "rb");
   if (!fp) {
-	fprintf(stderr, "Error: could not open file '%s'\n", filename);
-	return NULL;
+    fprintf(stderr, "Error: could not open file '%s'\n", filename);
+    return NULL;
   }
 
   // Read header (magic number, width, height, and max color value)
@@ -73,24 +60,24 @@ PpmImage *ReadPpm(const char *filename) {
   uint32_t height;
   uint16_t max_color;
   if (fscanf(fp, "%2s%*[ \t\r\n]%u%*[ \t\r\n]%u%*[ \t\r\n]%hu%*1[ \t\r\n]", magic, &width, &height, &max_color) !=
-	  4) {
-	fprintf(stderr, "Error: invalid header in file '%s'\n", filename);
-	fclose(fp);
-	return NULL;
+      4) {
+    fprintf(stderr, "Error: invalid header in file '%s'\n", filename);
+    fclose(fp);
+    return NULL;
   }
 
   // Make sure the magic number is "P6" (binary PPM format)
   if (magic[0] != 'P' || magic[1] != '6') {
-	fprintf(stderr, "Error: unsupported file format in file '%s'\n", filename);
-	fclose(fp);
-	return NULL;
+    fprintf(stderr, "Error: unsupported file format in file '%s'\n", filename);
+    fclose(fp);
+    return NULL;
   }
 
   // Make sure the max color value is PPM_MAX_COLOR
   if (max_color != PPM_MAX_COLOR) {
-	fprintf(stderr, "Error: max color value must be PPM_MAX_COLOR\n");
-	fclose(fp);
-	return NULL;
+    fprintf(stderr, "Error: max color value must be PPM_MAX_COLOR\n");
+    fclose(fp);
+    return NULL;
   }
 
   // Allocate memory for image data
@@ -99,11 +86,11 @@ PpmImage *ReadPpm(const char *filename) {
 
   // Read pixel data
   if (fread(image->data_, sizeof(Pixel), width * height, fp) != width * height) {
-	fprintf(stderr, "Error: could not read pixel data from file '%s'\n", filename);
-	free(image->data_);
-	free(image);
-	fclose(fp);
-	return NULL;
+    fprintf(stderr, "Error: could not read pixel data from file '%s'\n", filename);
+    free(image->data_);
+    free(image);
+    fclose(fp);
+    return NULL;
   }
 
   fclose(fp);
@@ -136,9 +123,9 @@ double SRgbValue(double linear_rgb) {
  * @param s_rgb SRgb Pixel
  */
 void LinearRgb(Pixel *s_rgb) {
-  s_rgb->r_ = (uint8_t)(LinearRgbValue(s_rgb->r_ / PPM_MAX_COLOR_F) * PPM_MAX_COLOR_F);
-  s_rgb->g_ = (uint8_t)(LinearRgbValue(s_rgb->g_ / PPM_MAX_COLOR_F) * PPM_MAX_COLOR_F);
-  s_rgb->b_ = (uint8_t)(LinearRgbValue(s_rgb->b_ / PPM_MAX_COLOR_F) * PPM_MAX_COLOR_F);
+  s_rgb->r_ = (uint8_t) (LinearRgbValue(s_rgb->r_ / PPM_MAX_COLOR_F) * PPM_MAX_COLOR_F);
+  s_rgb->g_ = (uint8_t) (LinearRgbValue(s_rgb->g_ / PPM_MAX_COLOR_F) * PPM_MAX_COLOR_F);
+  s_rgb->b_ = (uint8_t) (LinearRgbValue(s_rgb->b_ / PPM_MAX_COLOR_F) * PPM_MAX_COLOR_F);
 }
 
 /**
@@ -147,9 +134,9 @@ void LinearRgb(Pixel *s_rgb) {
  * @param linear_rgb Linear RGB color
  */
 void SRgb(Pixel *linear_rgb) {
-  linear_rgb->r_ = (uint8_t)(SRgbValue(linear_rgb->r_ / PPM_MAX_COLOR_F) * PPM_MAX_COLOR_F);
-  linear_rgb->g_ = (uint8_t)(SRgbValue(linear_rgb->g_ / PPM_MAX_COLOR_F) * PPM_MAX_COLOR_F);
-  linear_rgb->b_ = (uint8_t)(SRgbValue(linear_rgb->b_ / PPM_MAX_COLOR_F) * PPM_MAX_COLOR_F);
+  linear_rgb->r_ = (uint8_t) (SRgbValue(linear_rgb->r_ / PPM_MAX_COLOR_F) * PPM_MAX_COLOR_F);
+  linear_rgb->g_ = (uint8_t) (SRgbValue(linear_rgb->g_ / PPM_MAX_COLOR_F) * PPM_MAX_COLOR_F);
+  linear_rgb->b_ = (uint8_t) (SRgbValue(linear_rgb->b_ / PPM_MAX_COLOR_F) * PPM_MAX_COLOR_F);
 }
 
 /**
@@ -163,8 +150,8 @@ PpmImage *PpmPixelConvert(PpmImage *image, void (*conversion_fn)(Pixel *)) {
   // Allocate memory for the new image
   PpmImage *new_image = AllocatePpm(image->width_, image->height_);
   if (!new_image) {
-	fprintf(stderr, "Error: could not allocate memory for new PPM image");
-	return NULL;
+    fprintf(stderr, "Error: could not allocate memory for new PPM image");
+    return NULL;
   }
 
   // Copy the pixel data of the original image to the new image
@@ -173,7 +160,7 @@ PpmImage *PpmPixelConvert(PpmImage *image, void (*conversion_fn)(Pixel *)) {
   // Convert the pixel data of the new image using the given conversion function
 #pragma omp parallel for default(none) shared(new_image, conversion_fn)
   for (int i = 0; i < new_image->width_ * new_image->height_; i++)
-	conversion_fn(&new_image->data_[i]);
+    conversion_fn(&new_image->data_[i]);
 
   return new_image;
 }
@@ -199,37 +186,6 @@ double SRgbLuminance(const Pixel *p) {
 }
 
 /**
- * Convert an image to a new image using the given pixel conversion function.
- *
- * @param image     Pointer to the original image
- * @param luminance Reference to the luminance function
- * @return          Pointer to the new image
- */
-PgmImage *PpmToPgm(const PpmImage *image, double (*luminance)(const Pixel *)) {
-  /* Allocate memory for PGM image */
-  PgmImage *pgm_image = AllocatePgm(image->width_, image->height_);
-  if (!pgm_image) {
-	fprintf(stderr, "Error: could not allocate memory for PGM image\n");
-	return NULL;
-  }
-
-  // Convert pixel data from PPM image to PGM image
-#pragma omp parallel for default(none) shared(pgm_image, image, luminance)
-  for (uint32_t i = 0; i < pgm_image->height_ * pgm_image->width_; i++) {
-	// Get pixel from PPM image
-	Pixel p = image->data_[i];
-
-	// Calculate LinearLuminance of pixel
-	uint8_t y = (uint8_t)luminance(&p);
-
-	// Set LinearLuminance value in PGM image
-	pgm_image->data_[i] = y;
-  }
-
-  return pgm_image;
-}
-
-/**
  * Write a PPM image to a file.
  *
  * @param filename  The name of the file to write.
@@ -240,23 +196,23 @@ bool WritePpm(const char *filename, const PpmImage *image) {
   // Open file for writing
   FILE *fp = fopen(filename, "wb");
   if (!fp) {
-	fprintf(stderr, "Error: could not open file '%s' for writing\n", filename);
-	return false;
+    fprintf(stderr, "Error: could not open file '%s' for writing\n", filename);
+    return false;
   }
 
   // Write magic number, width, height, and max color value
   if (fprintf(fp, "P6\n%u\n%u\n%hu\n", image->width_, image->height_, image->max_color_) < 0) {
-	fprintf(stderr, "Error: could not write header to file '%s'\n", filename);
-	fclose(fp);
-	return false;
+    fprintf(stderr, "Error: could not write header to file '%s'\n", filename);
+    fclose(fp);
+    return false;
   }
 
   // Write pixel data
   if (fwrite(image->data_, sizeof(uint8_t), image->width_ * image->height_ * 3, fp) !=
-	  image->width_ * image->height_ * 3) {
-	fprintf(stderr, "Error: could not write pixel data to file '%s'\n", filename);
-	fclose(fp);
-	return false;
+      image->width_ * image->height_ * 3) {
+    fprintf(stderr, "Error: could not write pixel data to file '%s'\n", filename);
+    fclose(fp);
+    return false;
   }
 
   fclose(fp);
