+++ types/pixel.h
@@ -1,0 +1,13 @@
+#ifndef NETPBM_TYPES_PIXEL_H_
+#define NETPBM_TYPES_PIXEL_H_
+
+#include <stdint.h>
+
+// Pixel struct
+typedef struct {
+  uint8_t r_;
+  uint8_t g_;
+  uint8_t b_;
+} Pixel;
+
+#endif //NETPBM_TYPES_PIXEL_H_
